Implement `ConvexTriangleMesh` and `ConvexHull3d`
