+++ examples/shapes_3d.rs
@@ -10,7 +10,7 @@ use bevy::{
     prelude::*,
     render::render_resource::{Extent3d, TextureDimension, TextureFormat},
 };
-use quickhull::ConvexHull3d;
+use quickhull::ConvexTriangleMesh;
 
 fn main() {
     App::new()
@@ -197,7 +197,7 @@ fn render_convex_hulls(
             .collect::<Vec<_>>();
 
         let Ok(hull) =
-            ConvexHull3d::try_from_points(&positions, None).map_err(|e| warn!("{:?}", e))
+            ConvexTriangleMesh::try_from_points(&positions, None).map_err(|e| warn!("{:?}", e))
         else {
             continue;
         };
