+++ src/lib.rs
@@ -18,4 +18,6 @@ mod dim2;
 mod dim3;
 
 pub use dim2::ConvexHull2d;
-pub use dim3::{ConvexHull3d, ConvexHull3dError};
+pub use dim3::{
+    ConvexHull3d, ConvexHull3dError, ConvexHull3dSettings, ConvexTriangleMesh, HullFace, Plane3d,
+};
